Try a different approach to getting project info.
`ProjectInfoService` was firing a timer every second to check for changes to the solution. Because `DTE` must be queried on the UI thread, we were introducing slight pauses to the VS UI every second.

Try a different approach: what we really want to be able to say is "give me a list of all projects in the solution when they are fully loaded". The problem is there doesn't seem to be a way to tell whether all projects in the solution are fully loaded.

The `SolutionService` class introduced here tries to achieve that by assuming that a C# project with 0 references is not yet fully loaded (because there should at least be references to the framework). When such a situation is detected, it attaches event handlers to listen for references being loaded for each project and waits for all projects to have references.