Fix for #190 - taking assembly from debug rather than active configuration
This is my first attempt at fixing #190

I'm not expecting it to be merged (please don't), but rather I'm looking for an opinion for if I'm heading in the correct direction.

I'm thinking that if hooking into IVsUpdateSolutionEvents is correct, then perhaps SolutionService should implement it so I can remove UpdateSolutionEvents.